Resolve linting errors as needed
This commit attempts to resolve the linting errors, mostly by fixing the
code for readability and/or correctness as per ESLint's rules as
defined.  I changed one rule to a warning, though:
`no-warning-comments`.  I don't know why we're nulling out the
parameters passed in that one function, so I tagged it as such.  If an
explanation can be given as to why it's necessary to do so, I'll gladly
remove my objecting comment.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>