Introduce `@babel` to the mix as a proposal
Babel makes it so that you don't have to worry about what code is or
isn't supported by target; it started out as the project `6to5`, meaning
people could use it to write ES6 code when it was not at all supported.
It's through these efforts that modern JS has really flourished, in my
opinion.  Most features can even still be defined in JS 5, but this is
just some syntactic sugar.

The purpose of this branch is to show what it would look like using
babel 7 to build the project.  To that end, I'm also bringing in a few
eslint helpers, covered in the next commit.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>