Three point zero
I can't think of any other chores to do, so I guess we're doing it live.

Assuming everything is configured correctly, the next step is to build
and publish.

I'd recommend for running Coveralls locally, leave the `.coveralls.yml`
file as-is and adding in a `.env` file to keep the repo token.  It's
already ignored, and the repo token can be added within Travis for that
part.

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>